+++ pkg/ent/schema/dir.go
@@ -16,7 +16,7 @@ func (Dir) Fields() []ent.Field {
 		field.String("id"),
 		field.String("sync_id"),
 		field.String("dir"),
-		field.Uint64("level"),
+		field.Int("level"),
 		field.Bool("deleted"),
 		field.Int64("create_time"),
 		field.Int64("mod_time"),
